requirements.txt
flask==3.0.0
flask-cors==4.0.0
mysql-connector-python==8.3.0
bcrypt==4.1.2
PyJWT==2.8.0
requests==2.31.0